+++ sdk/cosmos/azure-cosmos-kafka-connect/CHANGELOG.md
@@ -7,7 +7,7 @@
 #### Breaking Changes
 
 #### Bugs Fixed
-* Source Connector status appears as `FAILED` if try to create a connector with an incorrect container name - See [PR 41160](https://github.com/Azure/azure-sdk-for-java/pull/41160) 
+* Connector status appears as `FAILED` if try to create a connector with an incorrect container names - See [PR 41160](https://github.com/Azure/azure-sdk-for-java/pull/41160) 
 
 #### Other Changes